+++ app/Http/Controllers/EmprestimoController.php
@@ -10,7 +10,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Session;
 
-class EmprestimosController extends Controller
+class EmprestimoController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -49,7 +49,7 @@ class EmprestimosController extends Controller
      */
     public function create()
     {
-        if (Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
             $contatos = Contato::all();
             $livros = Livro::all();
             return view('emprestimo.create',['contatos'=>$contatos,'livros'=>$livros]);
@@ -66,7 +66,7 @@ class EmprestimosController extends Controller
      */
     public function store(Request $request)
     {
-        if (Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
             $this->validate($request,[
                 'contato_id' => 'required',
                 'livro_id' => 'required',
@@ -121,7 +121,7 @@ class EmprestimosController extends Controller
      */
     public function devolver(Request $request, $id)
     {
-        if (Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
             $emprestimo = Emprestimo::find($id);
             $emprestimo->datadevolucao = \Carbon\Carbon::now();
             $emprestimo->save();
@@ -155,7 +155,7 @@ class EmprestimosController extends Controller
      */
     public function destroy(Request $request, $id)
     {
-        if (Auth::check()) {
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
             $emprestimo = Emprestimo::find($id);
 
             $emprestimo->delete();
